+++ lib/get_stream_page.dart
@@ -1,5 +1,6 @@
 library get_stream_page;
 
 export 'src/ui/get_stream_page.dart';
+export 'src/ui/get_stream_widget.dart';
 export 'src/ui/utils/functions.dart';
 export 'src/ui/infra/rx_get_type.dart';