+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## [0.0.1]
+## [0.1.0]
+* Added ```[GetStreamWidget]```  to build your body independently.
 
+## [0.0.2]
+* The readme with the example is very accessible.
+
+## [0.0.1]
 * Initial commit